requirements.txt
flask==2.3.3
boto3==1.28.44
python-dotenv==1.0.0
Flask-WTF==1.1.1
Werkzeug==2.3.7
bcrypt==4.0.1
Flask-Login==0.6.2
colorama==0.4.6